Agregada pagina animal, correcciones menores

+++ src/main/java/co/edu/konrad/zoologico/persistence/AnimalPersistance.java
@@ -24,7 +24,7 @@ public class AnimalPersistance {
     */
     @PersistenceContext(unitName = "zoologicoPU")
     protected EntityManager em;
-    
+
     /**
     * Metodo para insertar a los animales
     * @param a Objeto tipo AnimalEntity
@@ -34,7 +34,7 @@ public class AnimalPersistance {
         em.persist(a);
         return a;
     }
-    
+
     /**
     * Metodo para listar todos los animales
     * @return Devuelve una lista con todos los objetos de tipo AnimalEntity
@@ -43,29 +43,29 @@ public class AnimalPersistance {
         Query q = em.createQuery("select ta from AnimalEntity ta");
         return q.getResultList();
     }
-    
+
     /**
     * Metodo para encontrar segun id
     * @param id Indice para buscar
     * @return Retorna un objeto de tipo AnimalEntity
     */
-    public AnimalEntity find(int id){
+    public AnimalEntity find(long id){
         return em.find(AnimalEntity.class, id);
     }
-    
+
     /**
     * Metodo para actualizar los datos de un animal
-    * @param ta Objeto tipo AnimalEntity 
+    * @param ta Objeto tipo AnimalEntity
     * @return Objeto AnimalEntity actualizado
     */
     public AnimalEntity update(AnimalEntity ta){
         return em.merge(ta);
     }
-    
+
     /**
     * Metodo para eliminar un animal
     * @param id Indice para buscar
-    */    
+    */
     public void delete(int id){
         AnimalEntity entity = find(id);
         em.remove(entity);
